function_service refactor stash
